extract classes from init
update tests; fix inconsistent indentation
