+++ tree_maker/config.py
@@ -1,0 +1,3 @@
+DEFAULT_EXCLUDED_FILES = [
+  '__pycache__', 'venv', 'node_modules', 'dist', 'build', 'static', 'media'
+]
