+++ tree_maker/utils.py
@@ -1,0 +1,2 @@
+def is_hidden(item):
+  return item.startswith('.')
